+++ django_zero_downtime_migrations/backends/postgres/schema.py
@@ -8,6 +8,7 @@ from django.db.backends.ddl_references import Statement
 from django.db.backends.postgresql.schema import (
     DatabaseSchemaEditor as PostgresDatabaseSchemaEditor
 )
+from django.utils.functional import cached_property
 
 
 class Unsafe:
@@ -56,6 +57,17 @@ class UnsafeOperationException(Exception):
     pass
 
 
+class DummySQL:
+    def __mod__(self, other):
+        return DUMMY_SQL
+
+    def format(self, *args, **kwargs):
+        return DUMMY_SQL
+
+
+DUMMY_SQL = DummySQL()
+
+
 class MultiStatementSQL(list):
 
     def __init__(self, obj, *args):
@@ -70,9 +82,17 @@ class MultiStatementSQL(list):
         return str(self)
 
     def __mod__(self, other):
+        if other is DUMMY_SQL:
+            return DUMMY_SQL
+        if isinstance(other, (list, tuple)) and any(arg is DUMMY_SQL for arg in other):
+            return DUMMY_SQL
+        if isinstance(other, dict) and any(val is DUMMY_SQL for val in other.values()):
+            return DUMMY_SQL
         return MultiStatementSQL(s % other for s in self)
 
     def format(self, *args, **kwargs):
+        if any(arg is DUMMY_SQL for arg in args) or any(val is DUMMY_SQL for val in kwargs.values()):
+            return DUMMY_SQL
         return MultiStatementSQL(s.format(*args, **kwargs) for s in self)
 
 
@@ -92,9 +112,17 @@ class PGLock:
         return str(self)
 
     def __mod__(self, other):
+        if other is DUMMY_SQL:
+            return DUMMY_SQL
+        if isinstance(other, (list, tuple)) and any(arg is DUMMY_SQL for arg in other):
+            return DUMMY_SQL
+        if isinstance(other, dict) and any(val is DUMMY_SQL for val in other.values()):
+            return DUMMY_SQL
         return self.__class__(self.sql % other, self.use_timeouts, self.disable_statement_timeout)
 
     def format(self, *args, **kwargs):
+        if any(arg is DUMMY_SQL for arg in args) or any(val is DUMMY_SQL for val in kwargs.values()):
+            return DUMMY_SQL
         return self.__class__(self.sql.format(*args, **kwargs), self.use_timeouts, self.disable_statement_timeout)
 
 
@@ -110,6 +138,7 @@ class PGShareUpdateExclusive(PGLock):
 
 class DatabaseSchemaEditorMixin:
     ZERO_TIMEOUT = '0ms'
+    USE_PG_ATTRIBUTE_UPDATE_FOR_SUPERUSER = 'USE_PG_ATTRIBUTE_UPDATE_FOR_SUPERUSER'
 
     sql_get_lock_timeout = "SELECT setting || unit FROM pg_settings WHERE name = 'lock_timeout'"
     sql_get_statement_timeout = "SELECT setting || unit FROM pg_settings WHERE name = 'statement_timeout'"
@@ -185,27 +214,64 @@ class DatabaseSchemaEditorMixin:
     _sql_table_count = "SELECT reltuples FROM pg_class WHERE oid = '%(table)s'::regclass"
     _sql_check_notnull_constraint = (
         "SELECT conname FROM pg_constraint "
-        "WHERE contype = 'c' AND conrelid = '%(table)s'::regclass AND consrc = '(%(columns)s IS NOT NULL)'"
+        "WHERE contype = 'c' "
+        "AND conrelid = '%(table)s'::regclass "
+        "AND conname LIKE '%%_notnull'"
+        "AND pg_get_constraintdef(oid) = replace('CHECK ((%(columns)s IS NOT NULL))', '\"', '')"
     )
-    _sql_column_not_null_compatible = MultiStatementSQL(
+    _sql_column_not_null_compatible_le_pg12 = MultiStatementSQL(
         PGAccessExclusive("ALTER TABLE %(table)s ADD CONSTRAINT %(name)s CHECK (%(column)s IS NOT NULL) NOT VALID"),
         PGShareUpdateExclusive("ALTER TABLE %(table)s VALIDATE CONSTRAINT %(name)s",
                                disable_statement_timeout=True),
+    )
+    _sql_column_not_null_le_pg12_pg_attributes_for_root = MultiStatementSQL(
+        *_sql_column_not_null_compatible_le_pg12,
+        # pg_catalog.pg_attribute update require extra privileges
+        # that can be granted manually of already available for superusers
+        "UPDATE pg_catalog.pg_attribute SET attnotnull = TRUE "
+        "WHERE attrelid = '%(table)s'::regclass::oid AND attname = replace('%(column)s', '\"', '')",
+        PGAccessExclusive("ALTER TABLE %(table)s DROP CONSTRAINT %(name)s"),
+    )
+
+    _sql_column_not_null = MultiStatementSQL(
+        *_sql_column_not_null_compatible_le_pg12,
+        PGAccessExclusive("ALTER TABLE %(table)s ALTER COLUMN %(column)s SET NOT NULL"),
+        PGAccessExclusive("ALTER TABLE %(table)s DROP CONSTRAINT %(name)s"),
     )
 
     _varchar_type_regexp = re.compile(r'^varchar\((?P<max_length>\d+)\)$')
     _numeric_type_regexp = re.compile(r'^numeric\((?P<precision>\d+), *(?P<scale>\d+)\)$')
 
+    @cached_property
+    def is_postgresql_12(self):
+        return self.connection.pg_version >= 120000
+
     def __init__(self, connection, collect_sql=False, atomic=True):
+        # Disable atomic transactions as it can be reason of downtime or deadlock
+        # in case if you combine many operation in one migration module.
+        super().__init__(connection, collect_sql=collect_sql, atomic=False)
+
+        # Avoid using DUMMY_SQL in combined alters
+        connection.features.supports_combined_alters = False
+
+        # Get settings with defaults
         self.LOCK_TIMEOUT = getattr(settings, "ZERO_DOWNTIME_MIGRATIONS_LOCK_TIMEOUT", None)
         self.STATEMENT_TIMEOUT = getattr(settings, "ZERO_DOWNTIME_MIGRATIONS_STATEMENT_TIMEOUT", None)
         self.FLEXIBLE_STATEMENT_TIMEOUT = getattr(
             settings, "ZERO_DOWNTIME_MIGRATIONS_FLEXIBLE_STATEMENT_TIMEOUT", False)
+        if self.is_postgresql_12 and hasattr(settings, "ZERO_DOWNTIME_MIGRATIONS_USE_NOT_NULL"):
+            warnings.warn(
+                'settings.ZERO_DOWNTIME_MIGRATIONS_USE_NOT_NULL not applicable for postgres 12+. '
+                'Please remove this setting. If you migrated form old version, please move to NOT NULL constraint '
+                'from CHECK IS NOT NULL before with `migrate_isnotnull_check_constraints` management command.',
+                DeprecationWarning
+            )
         self.USE_NOT_NULL = getattr(settings, "ZERO_DOWNTIME_MIGRATIONS_USE_NOT_NULL", None)
         self.RAISE_FOR_UNSAFE = getattr(settings, "ZERO_DOWNTIME_MIGRATIONS_RAISE_FOR_UNSAFE", False)
-        super().__init__(connection, collect_sql=collect_sql, atomic=False)
 
     def execute(self, sql, params=()):
+        if sql is DUMMY_SQL:
+            return
         statements = []
         if isinstance(sql, MultiStatementSQL):
             statements.extend(sql)
@@ -256,19 +322,73 @@ class DatabaseSchemaEditorMixin:
         if lock_timeout is not None:
             self.execute(self.sql_set_lock_timeout % {"lock_timeout": previous_lock_timeout})
 
+    def _flush_deferred_sql(self):
+        """As some alternative sql use deferred sql and deferred sql run after all operations in miration module
+         so good idea to run deferred sql as soon as possible to provide similar as possible state
+         between operations in migration module."""
+        for sql in self.deferred_sql:
+            self.execute(sql)
+        self.deferred_sql.clear()
+
+    def create_model(self, model):
+        super().create_model(model)
+        self._flush_deferred_sql()
+
+    def delete_model(self, model):
+        super().delete_model(model)
+        self._flush_deferred_sql()
+
+    def alter_index_together(self, model, old_index_together, new_index_together):
+        super().alter_index_together(model, old_index_together, new_index_together)
+        self._flush_deferred_sql()
+
+    def alter_unique_together(self, model, old_unique_together, new_unique_together):
+        super().alter_unique_together(model, old_unique_together, new_unique_together)
+        self._flush_deferred_sql()
+
+    def add_index(self, model, index):
+        super().add_index(model, index)
+        self._flush_deferred_sql()
+
+    def remove_index(self, model, index):
+        super().remove_index(model, index)
+        self._flush_deferred_sql()
+
+    def add_constraint(self, model, constraint):
+        super().add_constraint(model, constraint)
+        self._flush_deferred_sql()
+
+    def remove_constraint(self, model, constraint):
+        super().remove_constraint(model, constraint)
+        self._flush_deferred_sql()
+
+    def add_field(self, model, field):
+        super().add_field(model, field)
+        self._flush_deferred_sql()
+
+    def remove_field(self, model, field):
+        super().remove_field(model, field)
+        self._flush_deferred_sql()
+
+    def alter_field(self, model, old_field, new_field, strict=False):
+        super().alter_field(model, old_field, new_field, strict)
+        self._flush_deferred_sql()
+
     def alter_db_table(self, model, old_db_table, new_db_table):
         if self.RAISE_FOR_UNSAFE:
             raise UnsafeOperationException(Unsafe.ALTER_TABLE_RENAME)
         else:
             warnings.warn(UnsafeOperationWarning(Unsafe.ALTER_TABLE_RENAME))
-        return super().alter_db_table(model, old_db_table, new_db_table)
+        super().alter_db_table(model, old_db_table, new_db_table)
+        self._flush_deferred_sql()
 
     def alter_db_tablespace(self, model, old_db_tablespace, new_db_tablespace):
         if self.RAISE_FOR_UNSAFE:
             raise UnsafeOperationException(Unsafe.ALTER_TABLE_SET_TABLESPACE)
         else:
             warnings.warn(UnsafeOperationWarning(Unsafe.ALTER_TABLE_SET_TABLESPACE))
-        return super().alter_db_tablespace(model, old_db_tablespace, new_db_tablespace)
+        super().alter_db_tablespace(model, old_db_tablespace, new_db_tablespace)
+        self._flush_deferred_sql()
 
     def _rename_field_sql(self, table, old_field, new_field, new_type):
         if self.RAISE_FOR_UNSAFE:
@@ -283,6 +403,9 @@ class DatabaseSchemaEditorMixin:
             cursor.execute(sql)
             rows_count, = cursor.fetchone()
         return rows_count
+
+    def _use_pg_attribute_update_for_not_null(self):
+        return self.USE_NOT_NULL == self.USE_PG_ATTRIBUTE_UPDATE_FOR_SUPERUSER
 
     def _use_check_constraint_for_not_null(self, model):
         if self.USE_NOT_NULL is True:
@@ -303,18 +426,11 @@ class DatabaseSchemaEditorMixin:
         return " DEFAULT %s"
 
     def _add_column_not_null(self, model, field):
-        if self.RAISE_FOR_UNSAFE and self.USE_NOT_NULL is None:
+        if self.RAISE_FOR_UNSAFE:
             raise UnsafeOperationException(Unsafe.ADD_COLUMN_NOT_NULL)
-        if self._use_check_constraint_for_not_null(model):
-            self.deferred_sql.append(self._sql_column_not_null_compatible % {
-                "column": self.quote_name(field.column),
-                "table": self.quote_name(model._meta.db_table),
-                "name": self.quote_name("{}_{}_notnull".format(model._meta.db_table, field.column)),
-            })
-            return ""
         else:
             warnings.warn(UnsafeOperationWarning(Unsafe.ADD_COLUMN_NOT_NULL))
-            return " NOT NULL"
+        return " NOT NULL"
 
     def _add_column_primary_key(self, model, field):
         self.deferred_sql.append(self.sql_create_pk % {
@@ -374,17 +490,35 @@ class DatabaseSchemaEditorMixin:
         return sql, params
 
     def _alter_column_set_not_null(self, model, new_field):
-        if self.RAISE_FOR_UNSAFE and self.USE_NOT_NULL is None:
+        if not self.is_postgresql_12 and self.RAISE_FOR_UNSAFE and self.USE_NOT_NULL is None:
             raise UnsafeOperationException(Unsafe.ALTER_COLUMN_NOT_NULL)
-        if self._use_check_constraint_for_not_null(model):
-            self.deferred_sql.append(self._sql_column_not_null_compatible % {
+        elif self.is_postgresql_12:
+            self.deferred_sql.append(self._sql_column_not_null % {
                 "column": self.quote_name(new_field.column),
                 "table": self.quote_name(model._meta.db_table),
                 "name": self.quote_name(
                     self._create_index_name(model._meta.db_table, [new_field.column], suffix="_notnull")
-                )
+                ),
             })
-            return None
+            return DUMMY_SQL, []
+        elif self._use_pg_attribute_update_for_not_null():
+            self.deferred_sql.append(self._sql_column_not_null_le_pg12_pg_attributes_for_root % {
+                "column": self.quote_name(new_field.column),
+                "table": self.quote_name(model._meta.db_table),
+                "name": self.quote_name(
+                    self._create_index_name(model._meta.db_table, [new_field.column], suffix="_notnull")
+                ),
+            })
+            return DUMMY_SQL, []
+        elif self._use_check_constraint_for_not_null(model):
+            self.deferred_sql.append(self._sql_column_not_null_compatible_le_pg12 % {
+                "column": self.quote_name(new_field.column),
+                "table": self.quote_name(model._meta.db_table),
+                "name": self.quote_name(
+                    self._create_index_name(model._meta.db_table, [new_field.column], suffix="_notnull")
+                ),
+            })
+            return DUMMY_SQL, []
         else:
             warnings.warn(UnsafeOperationWarning(Unsafe.ALTER_COLUMN_NOT_NULL))
             return self.sql_alter_column_not_null % {