+++ tests/apps/good_flow_app/models.py
@@ -1,0 +1,9 @@
+from django.db import models
+
+
+class TestTable(models.Model):
+    test_field = models.IntegerField()
+
+
+class RelatedTestTable(models.Model):
+    pass