+++ tests/unit/test_schema.py
@@ -1,4 +1,4 @@
-from functools import partial
+from functools import partial, wraps
 
 import django
 from django.conf import settings
@@ -10,13 +10,19 @@ from django.test import override_settings
 from django.utils.module_loading import import_string
 
 import pytest
+
 from django_zero_downtime_migrations.backends.postgres.schema import (
     UnsafeOperationException, UnsafeOperationWarning
 )
+from tests import skip_for_default_django_backend
+
+pytestmark = skip_for_default_django_backend
 
 DatabaseSchemaEditor = import_string(settings.DATABASES['default']['ENGINE'] + '.schema.DatabaseSchemaEditor')
 
 
+PG_VERSION_12 = 120000
+PG_VERSION_11 = 110000
 START_TIMEOUTS = [
     'SET statement_timeout TO \'0\';',
     'SET lock_timeout TO \'0\';',
@@ -45,6 +51,19 @@ def flexible_statement_timeout(statements):
     return START_FLEXIBLE_STATEMENT_TIMEOUT + statements + END_FLEXIBLE_STATEMENT_TIMEOUT
 
 
+def old_pg(func):
+    @wraps(func)
+    def wrapper(*args, **kwargs):
+        old = connection.pg_version
+        connection.pg_version = PG_VERSION_11
+        try:
+            result = func(*args, **kwargs)
+        finally:
+            connection.pg_version = old
+        return result
+    return wrapper
+
+
 class Model(models.Model):
     field1 = models.IntegerField()
     field2 = models.IntegerField()
@@ -54,6 +73,7 @@ class Model2(models.Model):
     pass
 
 
+connection.pg_version = PG_VERSION_12
 schema_editor = partial(DatabaseSchemaEditor, connection=connection, collect_sql=True)
 
 
@@ -180,8 +200,7 @@ def test_add_field_with_not_null__raise():
             editor.add_field(Model, field)
 
 
-@override_settings(ZERO_DOWNTIME_MIGRATIONS_RAISE_FOR_UNSAFE=True,
-                   ZERO_DOWNTIME_MIGRATIONS_USE_NOT_NULL=True)
+@override_settings(ZERO_DOWNTIME_MIGRATIONS_USE_NOT_NULL=True)
 def test_add_field_with_not_null__allowed_for_all_tables__warning():
     with cmp_schema_editor() as editor:
         with pytest.warns(UnsafeOperationWarning, match='ADD COLUMN NOT NULL is unsafe operation'):
@@ -191,8 +210,7 @@ def test_add_field_with_not_null__allowed_for_all_tables__warning():
     assert editor.collected_sql == timeouts(editor.core_editor.collected_sql)
 
 
-@override_settings(ZERO_DOWNTIME_MIGRATIONS_RAISE_FOR_UNSAFE=True,
-                   ZERO_DOWNTIME_MIGRATIONS_USE_NOT_NULL=10)
+@override_settings(ZERO_DOWNTIME_MIGRATIONS_USE_NOT_NULL=10)
 def test_add_field_with_not_null__allowed_for_small_tables__warning(mocker):
     mocker.patch.object(connection, 'cursor')().__enter__().fetchone.return_value = (5,)
     with cmp_schema_editor() as editor:
@@ -203,76 +221,90 @@ def test_add_field_with_not_null__allowed_for_small_tables__warning(mocker):
     assert editor.collected_sql == timeouts(editor.core_editor.collected_sql)
 
 
+@override_settings(ZERO_DOWNTIME_MIGRATIONS_USE_NOT_NULL=1)
+def test_add_field_with_not_null__use_compatible_constraint_for_large_tables__warning(mocker):
+    mocker.patch.object(connection, 'cursor')().__enter__().fetchone.return_value = (5,)
+    with cmp_schema_editor() as editor:
+        with pytest.warns(UnsafeOperationWarning, match='ADD COLUMN NOT NULL is unsafe operation'):
+            field = models.CharField(max_length=40, null=False)
+            field.set_attributes_from_name('field')
+            editor.add_field(Model, field)
+    assert editor.collected_sql == timeouts(editor.core_editor.collected_sql)
+
+
+@override_settings(ZERO_DOWNTIME_MIGRATIONS_USE_NOT_NULL=1,
+                   ZERO_DOWNTIME_MIGRATIONS_FLEXIBLE_STATEMENT_TIMEOUT=True)
+def test_add_field_with_not_null__use_compatible_constraint_for_large_tables__with_flexible_timeout__warning(mocker):
+    mocker.patch.object(connection, 'cursor')().__enter__().fetchone.return_value = (5,)
+    with cmp_schema_editor() as editor:
+        with pytest.warns(UnsafeOperationWarning, match='ADD COLUMN NOT NULL is unsafe operation'):
+            field = models.CharField(max_length=40, null=False)
+            field.set_attributes_from_name('field')
+            editor.add_field(Model, field)
+    assert editor.collected_sql == timeouts(editor.core_editor.collected_sql)
+
+
+@override_settings(ZERO_DOWNTIME_MIGRATIONS_USE_NOT_NULL=False)
+def test_add_field_with_not_null__use_compatible_constraint_for_all_tables__warning():
+    with cmp_schema_editor() as editor:
+        with pytest.warns(UnsafeOperationWarning, match='ADD COLUMN NOT NULL is unsafe operation'):
+            field = models.CharField(max_length=40, null=False)
+            field.set_attributes_from_name('field')
+            editor.add_field(Model, field)
+    assert editor.collected_sql == timeouts(editor.core_editor.collected_sql)
+
+
+@override_settings(ZERO_DOWNTIME_MIGRATIONS_USE_NOT_NULL=False,
+                   ZERO_DOWNTIME_MIGRATIONS_FLEXIBLE_STATEMENT_TIMEOUT=True)
+def test_add_field_with_not_null__use_compatible_constraint_for_all_tables__with_flexible_timeout__warning():
+    with cmp_schema_editor() as editor:
+        with pytest.warns(UnsafeOperationWarning, match='ADD COLUMN NOT NULL is unsafe operation'):
+            field = models.CharField(max_length=40, null=False)
+            field.set_attributes_from_name('field')
+            editor.add_field(Model, field)
+    assert editor.collected_sql == timeouts(editor.core_editor.collected_sql)
+
+
+@override_settings(ZERO_DOWNTIME_MIGRATIONS_RAISE_FOR_UNSAFE=True,
+                   ZERO_DOWNTIME_MIGRATIONS_USE_NOT_NULL=True)
+def test_add_field_with_not_null__allowed_for_all_tables__raise():
+    with cmp_schema_editor() as editor:
+        with pytest.raises(UnsafeOperationException, match='ADD COLUMN NOT NULL is unsafe operation'):
+            field = models.CharField(max_length=40, null=False)
+            field.set_attributes_from_name('field')
+            editor.add_field(Model, field)
+
+
+@override_settings(ZERO_DOWNTIME_MIGRATIONS_RAISE_FOR_UNSAFE=True,
+                   ZERO_DOWNTIME_MIGRATIONS_USE_NOT_NULL=10)
+def test_add_field_with_not_null__allowed_for_small_tables__raise(mocker):
+    mocker.patch.object(connection, 'cursor')().__enter__().fetchone.return_value = (5,)
+    with cmp_schema_editor() as editor:
+        with pytest.raises(UnsafeOperationException, match='ADD COLUMN NOT NULL is unsafe operation'):
+            field = models.CharField(max_length=40, null=False)
+            field.set_attributes_from_name('field')
+            editor.add_field(Model, field)
+
+
 @override_settings(ZERO_DOWNTIME_MIGRATIONS_RAISE_FOR_UNSAFE=True,
                    ZERO_DOWNTIME_MIGRATIONS_USE_NOT_NULL=1)
-def test_add_field_with_not_null__use_compatible_constraint_for_large_tables__ok(mocker):
+def test_add_field_with_not_null__use_compatible_constraint_for_large_tables__raise(mocker):
     mocker.patch.object(connection, 'cursor')().__enter__().fetchone.return_value = (5,)
     with cmp_schema_editor() as editor:
-        field = models.CharField(max_length=40, null=False)
-        field.set_attributes_from_name('field')
-        editor.add_field(Model, field)
-    assert editor.collected_sql == timeouts(
-        'ALTER TABLE "tests_model" ADD COLUMN "field" varchar(40);',
-    ) + timeouts(
-        'ALTER TABLE "tests_model" ADD CONSTRAINT "tests_model_field_notnull" '
-        'CHECK ("field" IS NOT NULL) NOT VALID;',
-    ) + [
-        'ALTER TABLE "tests_model" VALIDATE CONSTRAINT "tests_model_field_notnull";',
-    ]
-
-
-@override_settings(ZERO_DOWNTIME_MIGRATIONS_RAISE_FOR_UNSAFE=True,
-                   ZERO_DOWNTIME_MIGRATIONS_USE_NOT_NULL=1,
-                   ZERO_DOWNTIME_MIGRATIONS_FLEXIBLE_STATEMENT_TIMEOUT=True)
-def test_add_field_with_not_null__use_compatible_constraint_for_large_tables__with_flexible_timeout__ok(mocker):
-    mocker.patch.object(connection, 'cursor')().__enter__().fetchone.return_value = (5,)
-    with cmp_schema_editor() as editor:
-        field = models.CharField(max_length=40, null=False)
-        field.set_attributes_from_name('field')
-        editor.add_field(Model, field)
-    assert editor.collected_sql == timeouts(
-        'ALTER TABLE "tests_model" ADD COLUMN "field" varchar(40);',
-    ) + timeouts(
-        'ALTER TABLE "tests_model" ADD CONSTRAINT "tests_model_field_notnull" '
-        'CHECK ("field" IS NOT NULL) NOT VALID;',
-    ) + flexible_statement_timeout(
-        'ALTER TABLE "tests_model" VALIDATE CONSTRAINT "tests_model_field_notnull";',
-    )
+        with pytest.raises(UnsafeOperationException, match='ADD COLUMN NOT NULL is unsafe operation'):
+            field = models.CharField(max_length=40, null=False)
+            field.set_attributes_from_name('field')
+            editor.add_field(Model, field)
 
 
 @override_settings(ZERO_DOWNTIME_MIGRATIONS_RAISE_FOR_UNSAFE=True,
                    ZERO_DOWNTIME_MIGRATIONS_USE_NOT_NULL=False)
-def test_add_field_with_not_null__use_compatible_constraint_for_all_tables__ok():
+def test_add_field_with_not_null__use_compatible_constraint_for_all_tables__raise():
     with cmp_schema_editor() as editor:
-        field = models.CharField(max_length=40, null=False)
-        field.set_attributes_from_name('field')
-        editor.add_field(Model, field)
-    assert editor.collected_sql == timeouts(
-        'ALTER TABLE "tests_model" ADD COLUMN "field" varchar(40);',
-    ) + timeouts(
-        'ALTER TABLE "tests_model" ADD CONSTRAINT "tests_model_field_notnull" '
-        'CHECK ("field" IS NOT NULL) NOT VALID;',
-    ) + [
-        'ALTER TABLE "tests_model" VALIDATE CONSTRAINT "tests_model_field_notnull";',
-    ]
-
-
-@override_settings(ZERO_DOWNTIME_MIGRATIONS_RAISE_FOR_UNSAFE=True,
-                   ZERO_DOWNTIME_MIGRATIONS_USE_NOT_NULL=False,
-                   ZERO_DOWNTIME_MIGRATIONS_FLEXIBLE_STATEMENT_TIMEOUT=True)
-def test_add_field_with_not_null__use_compatible_constraint_for_all_tables__with_flexible_timeout__ok():
-    with cmp_schema_editor() as editor:
-        field = models.CharField(max_length=40, null=False)
-        field.set_attributes_from_name('field')
-        editor.add_field(Model, field)
-    assert editor.collected_sql == timeouts(
-        'ALTER TABLE "tests_model" ADD COLUMN "field" varchar(40);',
-    ) + timeouts(
-        'ALTER TABLE "tests_model" ADD CONSTRAINT "tests_model_field_notnull" '
-        'CHECK ("field" IS NOT NULL) NOT VALID;',
-    ) + flexible_statement_timeout(
-        'ALTER TABLE "tests_model" VALIDATE CONSTRAINT "tests_model_field_notnull";',
-    )
+        with pytest.raises(UnsafeOperationException, match='ADD COLUMN NOT NULL is unsafe operation'):
+            field = models.CharField(max_length=40, null=False)
+            field.set_attributes_from_name('field')
+            editor.add_field(Model, field)
 
 
 @override_settings(ZERO_DOWNTIME_MIGRATIONS_RAISE_FOR_UNSAFE=True)
@@ -535,7 +567,96 @@ def test_alter_field_decimal10_2_to_decimal10_1__raise():
             editor.alter_field(Model, old_field, new_field)
 
 
-def test_alter_field_set_not_null__warning():
+@override_settings(ZERO_DOWNTIME_MIGRATIONS_RAISE_FOR_UNSAFE=True)
+def test_alter_field_set_not_null__ok():
+    with cmp_schema_editor() as editor:
+        old_field = models.CharField(max_length=40, null=True)
+        old_field.set_attributes_from_name('field')
+        new_field = models.CharField(max_length=40, null=False)
+        new_field.set_attributes_from_name('field')
+        editor.alter_field(Model, old_field, new_field)
+    assert editor.collected_sql == timeouts(
+        'ALTER TABLE "tests_model" ADD CONSTRAINT "tests_model_field_0a53d95f_notnull" '
+        'CHECK ("field" IS NOT NULL) NOT VALID;',
+    ) + [
+        'ALTER TABLE "tests_model" VALIDATE CONSTRAINT "tests_model_field_0a53d95f_notnull";',
+    ] + timeouts(
+        'ALTER TABLE "tests_model" ALTER COLUMN "field" SET NOT NULL;'
+    ) + timeouts(
+        'ALTER TABLE "tests_model" DROP CONSTRAINT "tests_model_field_0a53d95f_notnull";'
+    )
+
+
+@override_settings(ZERO_DOWNTIME_MIGRATIONS_RAISE_FOR_UNSAFE=True,
+                   ZERO_DOWNTIME_MIGRATIONS_FLEXIBLE_STATEMENT_TIMEOUT=True)
+def test_alter_field_set_not_null__with_flexible_timeout__ok():
+    with cmp_schema_editor() as editor:
+        old_field = models.CharField(max_length=40, null=True)
+        old_field.set_attributes_from_name('field')
+        new_field = models.CharField(max_length=40, null=False)
+        new_field.set_attributes_from_name('field')
+        editor.alter_field(Model, old_field, new_field)
+    assert editor.collected_sql == timeouts(
+        'ALTER TABLE "tests_model" ADD CONSTRAINT "tests_model_field_0a53d95f_notnull" '
+        'CHECK ("field" IS NOT NULL) NOT VALID;',
+    ) + flexible_statement_timeout(
+        'ALTER TABLE "tests_model" VALIDATE CONSTRAINT "tests_model_field_0a53d95f_notnull";',
+    ) + timeouts(
+        'ALTER TABLE "tests_model" ALTER COLUMN "field" SET NOT NULL;'
+    ) + timeouts(
+        'ALTER TABLE "tests_model" DROP CONSTRAINT "tests_model_field_0a53d95f_notnull";'
+    )
+
+
+@override_settings(ZERO_DOWNTIME_MIGRATIONS_RAISE_FOR_UNSAFE=True,
+                   ZERO_DOWNTIME_MIGRATIONS_USE_NOT_NULL='USE_PG_ATTRIBUTE_UPDATE_FOR_SUPERUSER')
+@old_pg
+def test_alter_field_set_not_null__old_pg__use_pg_attribute_update__ok():
+    with cmp_schema_editor() as editor:
+        old_field = models.CharField(max_length=40, null=True)
+        old_field.set_attributes_from_name('field')
+        new_field = models.CharField(max_length=40, null=False)
+        new_field.set_attributes_from_name('field')
+        editor.alter_field(Model, old_field, new_field)
+    assert editor.collected_sql == timeouts(
+        'ALTER TABLE "tests_model" ADD CONSTRAINT "tests_model_field_0a53d95f_notnull" '
+        'CHECK ("field" IS NOT NULL) NOT VALID;',
+    ) + [
+        'ALTER TABLE "tests_model" VALIDATE CONSTRAINT "tests_model_field_0a53d95f_notnull";',
+    ] + [
+        'UPDATE pg_catalog.pg_attribute SET attnotnull = TRUE '
+        'WHERE attrelid = \'"tests_model"\'::regclass::oid AND attname = replace(\'"field"\', \'"\', \'\');',
+    ] + timeouts(
+        'ALTER TABLE "tests_model" DROP CONSTRAINT "tests_model_field_0a53d95f_notnull";'
+    )
+
+
+@override_settings(ZERO_DOWNTIME_MIGRATIONS_RAISE_FOR_UNSAFE=True,
+                   ZERO_DOWNTIME_MIGRATIONS_USE_NOT_NULL='USE_PG_ATTRIBUTE_UPDATE_FOR_SUPERUSER',
+                   ZERO_DOWNTIME_MIGRATIONS_FLEXIBLE_STATEMENT_TIMEOUT=True)
+@old_pg
+def test_alter_field_set_not_null__old_pg__use_pg_attribute_update__with_flexible_timeout__ok():
+    with cmp_schema_editor() as editor:
+        old_field = models.CharField(max_length=40, null=True)
+        old_field.set_attributes_from_name('field')
+        new_field = models.CharField(max_length=40, null=False)
+        new_field.set_attributes_from_name('field')
+        editor.alter_field(Model, old_field, new_field)
+    assert editor.collected_sql == timeouts(
+        'ALTER TABLE "tests_model" ADD CONSTRAINT "tests_model_field_0a53d95f_notnull" '
+        'CHECK ("field" IS NOT NULL) NOT VALID;',
+    ) + flexible_statement_timeout(
+        'ALTER TABLE "tests_model" VALIDATE CONSTRAINT "tests_model_field_0a53d95f_notnull";',
+    ) + [
+        'UPDATE pg_catalog.pg_attribute SET attnotnull = TRUE '
+        'WHERE attrelid = \'"tests_model"\'::regclass::oid AND attname = replace(\'"field"\', \'"\', \'\');',
+    ] + timeouts(
+        'ALTER TABLE "tests_model" DROP CONSTRAINT "tests_model_field_0a53d95f_notnull";'
+    )
+
+
+@old_pg
+def test_alter_field_set_not_null__old_pg__warning():
     with cmp_schema_editor() as editor:
         with pytest.warns(UnsafeOperationWarning, match='ALTER COLUMN NOT NULL is unsafe operation'):
             old_field = models.CharField(max_length=40, null=True)
@@ -547,7 +668,8 @@ def test_alter_field_set_not_null__warning():
 
 
 @override_settings(ZERO_DOWNTIME_MIGRATIONS_RAISE_FOR_UNSAFE=True)
-def test_alter_field_set_not_null__raise():
+@old_pg
+def test_alter_field_set_not_null__old_pg__raise():
     with cmp_schema_editor() as editor:
         with pytest.raises(UnsafeOperationException, match='ALTER COLUMN NOT NULL is unsafe operation'):
             old_field = models.CharField(max_length=40, null=True)
@@ -559,7 +681,8 @@ def test_alter_field_set_not_null__raise():
 
 @override_settings(ZERO_DOWNTIME_MIGRATIONS_RAISE_FOR_UNSAFE=True,
                    ZERO_DOWNTIME_MIGRATIONS_USE_NOT_NULL=True)
-def test_alter_field_set_not_null__allowed_for_all_tables__warning():
+@old_pg
+def test_alter_field_set_not_null__old_pg__allowed_for_all_tables__warning():
     with cmp_schema_editor() as editor:
         with pytest.warns(UnsafeOperationWarning, match='ALTER COLUMN NOT NULL is unsafe operation'):
             old_field = models.CharField(max_length=40, null=True)
@@ -572,7 +695,8 @@ def test_alter_field_set_not_null__allowed_for_all_tables__warning():
 
 @override_settings(ZERO_DOWNTIME_MIGRATIONS_RAISE_FOR_UNSAFE=True,
                    ZERO_DOWNTIME_MIGRATIONS_USE_NOT_NULL=10)
-def test_alter_field_set_not_null__allowed_for_small_tables__warning(mocker):
+@old_pg
+def test_alter_field_set_not_null__old_pg__allowed_for_small_tables__warning(mocker):
     mocker.patch.object(connection, 'cursor')().__enter__().fetchone.return_value = (5,)
     with cmp_schema_editor() as editor:
         with pytest.warns(UnsafeOperationWarning, match='ALTER COLUMN NOT NULL is unsafe operation'):
@@ -586,7 +710,8 @@ def test_alter_field_set_not_null__allowed_for_small_tables__warning(mocker):
 
 @override_settings(ZERO_DOWNTIME_MIGRATIONS_RAISE_FOR_UNSAFE=True,
                    ZERO_DOWNTIME_MIGRATIONS_USE_NOT_NULL=1)
-def test_alter_field_set_not_null__use_compatible_constraint_for_large_tables__ok(mocker):
+@old_pg
+def test_alter_field_set_not_null__old_pg__use_compatible_constraint_for_large_tables__ok(mocker):
     mocker.patch.object(connection, 'cursor')().__enter__().fetchone.return_value = (5,)
     with cmp_schema_editor() as editor:
         old_field = models.CharField(max_length=40, null=True)
@@ -605,7 +730,10 @@ def test_alter_field_set_not_null__use_compatible_constraint_for_large_tables__o
 @override_settings(ZERO_DOWNTIME_MIGRATIONS_RAISE_FOR_UNSAFE=True,
                    ZERO_DOWNTIME_MIGRATIONS_USE_NOT_NULL=1,
                    ZERO_DOWNTIME_MIGRATIONS_FLEXIBLE_STATEMENT_TIMEOUT=True)
-def test_alter_field_set_not_null__use_compatible_constraint_for_large_tables__with_flexible_timeout__ok(mocker):
+@old_pg
+def test_alter_field_set_not_null__old_pg__use_compatible_constraint_for_large_tables__with_flexible_timeout__ok(
+    mocker
+):
     mocker.patch.object(connection, 'cursor')().__enter__().fetchone.return_value = (5,)
     with cmp_schema_editor() as editor:
         old_field = models.CharField(max_length=40, null=True)
@@ -623,7 +751,8 @@ def test_alter_field_set_not_null__use_compatible_constraint_for_large_tables__w
 
 @override_settings(ZERO_DOWNTIME_MIGRATIONS_RAISE_FOR_UNSAFE=True,
                    ZERO_DOWNTIME_MIGRATIONS_USE_NOT_NULL=False)
-def test_alter_field_set_not_null__use_compatible_constraint_for_all_tables__ok():
+@old_pg
+def test_alter_field_set_not_null__old_pg__use_compatible_constraint_for_all_tables__ok():
     with cmp_schema_editor() as editor:
         old_field = models.CharField(max_length=40, null=True)
         old_field.set_attributes_from_name('field')
@@ -641,7 +770,8 @@ def test_alter_field_set_not_null__use_compatible_constraint_for_all_tables__ok(
 @override_settings(ZERO_DOWNTIME_MIGRATIONS_RAISE_FOR_UNSAFE=True,
                    ZERO_DOWNTIME_MIGRATIONS_USE_NOT_NULL=False,
                    ZERO_DOWNTIME_MIGRATIONS_FLEXIBLE_STATEMENT_TIMEOUT=True)
-def test_alter_field_set_not_null__use_compatible_constraint_for_all_tables__with_flexible_timeout__ok():
+@old_pg
+def test_alter_field_set_not_null__old_pg__use_compatible_constraint_for_all_tables__with_flexible_timeout__ok():
     with cmp_schema_editor() as editor:
         old_field = models.CharField(max_length=40, null=True)
         old_field.set_attributes_from_name('field')