+++ CHANGES.md
@@ -1,13 +1,21 @@
 # django-pg-zero-downtime-migrations changelog
 
 ## 0.6
-  - add version to package
-  - fix pypi README images links
+  - marked `ZERO_DOWNTIME_MIGRATIONS_USE_NOT_NULL` option deprecated for postgres 12+
+  - added management command for migration to real `NOT NULL` from `CHECK IS NOT NULL` constraint
+  - added integration tests for pg 12, pg 11 root, pg 11 compatible not null constraint, pg 11 standard not null constraint and pg 10, 9.6, 9.5, 9.4, postgis databases
+  - fixed compatible check not null constraint deletion and creation via pg_attribute bugs
+  - minimized side affect with deferred sql execution between operations in one migration module
+  - added postgres 12 safe `NOT NULL` constraint creation
+  - added safe `NOT NULL` constraint creation for extra permissions for `pg_catalog.pg_attribute` with `ZERO_DOWNTIME_MIGRATIONS_USE_NOT_NULL=USE_PG_ATTRIBUTE_UPDATE_FOR_SUPERUSER` option enabled
+  - marked `AddField` with `null=False` parameter and compatible `CHECK IS NOT NULL` constraint option as unsafe operation and avoid `ZERO_DOWNTIME_MIGRATIONS_USE_NOT_NULL` value in this case
+  - added version to package
+  - fixed pypi README images links
   - improved README
 
 ## 0.5
   - extracted zero-downtime-schema to mixin to allow use this logic with other backends
-  - moveed module from `django_zero_downtime_migrations_postgres_backend` to `django_zero_downtime_migrations.backends.postgres`
+  - moved module from `django_zero_downtime_migrations_postgres_backend` to `django_zero_downtime_migrations.backends.postgres`
   - marked `django_zero_downtime_migrations_postgres_backend` module as deprecated
   - added postgis backend support
   - improved README